indexer/pyproject.toml
[tool.poetry]
name = "indexer"
version = "0.1.0"
description = "Indexador de processos para o desafio de estágio de busca do Justarter"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
elasticsearch = "^8.15.0"
psycopg2-binary = "^2.9.9"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
httpx = "^0.27.0"
pytest = "^8.4.2"
pytest-cov = "^5.0.0"


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
